requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
openai>=1.40.0
python-docx>=1.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
pytest>=8.0.0
